chat-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
langchain
langchain-openai
openai